Last minute Panicgit commit -m

+++ Assets/_Scripts/GameManager.cs
@@ -2,12 +2,12 @@
 
 public class GameManager : MonoBehaviour {
 
-	public static int score;
+	public static float score;
 	public static bool[] solved = {false, false, false, false, false, false}; // 0 is placeholder
 
 	void Start () {
 		score = 0;
 		Time.timeScale = 0;
 	}
-	
+
 }
